final commit

+++ routes/ideas.js
@@ -58,19 +58,24 @@ router.post('/', async (req, res) => {
 // Update an Idea
 router.put('/:id', async (req, res) => {
     try {
-        const updatedIdea = await Idea.findByIdAndUpdate(
-            req.params.id,
-            {
-                $set: {
-                    text: req.body.text,
-                    tag: req.body.tag
+        const idea = await Idea.findById(req.params.id);
+        if (idea.username = req.body.username) {
+            const updatedIdea = await Idea.findByIdAndUpdate(
+                req.params.id,
+                {
+                    $set: {
+                        text: req.body.text,
+                        tag: req.body.tag
+                    }
+                },
+                {
+                    new: true
                 }
-            },
-            {
-                new: true
-            }
-        )
-        res.json({ success: true, data: updatedIdea })
+            )
+            return res.json({ success: true, data: updatedIdea })
+        }
+        // If user do not match
+        res.status(403).json({ success: false, error: 'You are not authorized to update' })
     } catch (error) {
         console.log(error);
         res.status(500).json({ success: false, error: 'Something went wrong' })
@@ -79,8 +84,16 @@ router.put('/:id', async (req, res) => {
 // Delete an Idea
 router.delete('/:id', async (req, res) => {
     try {
-        await Idea.findByIdAndDelete(req.params.id)
-        res.send({ success: true, data: {} })
+        const idea = await Idea.findById(req.params.id);
+        // match the usernames
+        console.log(idea.username);
+        console.log(req.body.username);
+        if (idea.username === req.body.username) {
+            await Idea.findByIdAndDelete(req.params.id)
+            return res.send({ success: true, data: {} })
+        }
+        // usernames do not match
+        res.status(403).json({ success: false, error: 'You are not authorized to delete' })
     } catch (error) {
         console.log(error);
         res.status(500).json({ success: false, error: 'Something went wrong' })
